# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Closy API",
    "version": "1.0.0",
    "description": "closy-api dev test document"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地開發伺服器"
    },
    {
      "url": "https://closy-api.onrender.com",
      "description": "Render dev"
    }
  ],
  "paths": {
...
    "/home/outfit": {
      "post": {
        "tags": [
          "Home"
        ],
        "summary": "生成虛擬穿搭圖片",
        "description": "根據使用者選擇的衣物，使用 AI 合成虛擬穿搭圖片，需要攜帶 Bearer Token",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "虛擬穿搭圖片生成成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "虛擬穿搭圖片生成成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "outfitImgUrl": {
                          "type": "string",
                          "example": "https://res.cloudinary.com/test/closy/users/outfits/abc123/outfit.jpg"
                        },
                        "occasion": {
                          "type": "string",
                          "example": "casual"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "參數錯誤（selectedItems 不可為空 / 請提供正確的場合）",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "selectedItems 不可為空"
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授權或 Token 錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "未提供 Token 或格式錯誤"
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "找不到相關資訊",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 404
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "找不到使用者"
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "伺服器錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 500
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "伺服器發生不可預期的錯誤"
                    },
                    "data": {
                      "type": "object",
                      "example": null
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "selectedItems"
                ],
                "properties": {
                  "selectedItems": {
                    "type": "array",
                    "description": "選擇的衣物列表，不可為空",
                    "items": {
                      "type": "object",
                      "required": [
                        "cloudImgUrl",
                        "category",
                        "name",
                        "brand"
                      ],
                      "properties": {
                        "cloudImgUrl": {
                          "type": "string",
                          "example": "https://res.cloudinary.com/test/image.jpg"
                        },
                        "category": {
                          "type": "string",
                          "enum": [
                            "top",
                            "bottom",
                            "outerwear",
                            "shoes",
                            "skirt",
                            "dress"
                          ],
                          "example": "top"
                        },
                        "name": {
                          "type": "string",
                          "example": "白色基本款 T-shirt"
                        },
                        "brand": {
                          "type": "string",
                          "example": "UNIQLO"
                        }
                      }
                    }
                  },
                  "occasion": {
                    "type": "string",
                    "description": "場合（必填）",
                    "enum": [
                      "socialGathering",
                      "campusCasual",
                      "businessCasual",
                      "professional"
                    ],
                    "example": "socialGathering"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/outfit/{id}": {
      "delete": {
        "tags": [
          "Outfit"
        ],
        "summary": "刪除穿搭",
        "description": "依穿搭 ID 刪除指定穿搭紀錄。",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "664f1a2b3c4d5e6f7a8b9c0d"
            },
            "description": "穿搭 ID"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "刪除成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "刪除成功"
                    },
                    "data": {
                      "type": "object",
                      "example": {}
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "請求錯誤 (未提供 ID 或刪除失敗)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "請提供穿搭 id / 刪除失敗"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "身分驗證失敗 (可能原因：未提供 Token、Token 格式錯誤、Token 已過期)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "未提供 Token 或格式錯誤 / 無效的 Token 格式 / 無效的憑證或憑證已過期，請重新登入"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "伺服器錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 500
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "內部伺服器錯誤或其他錯誤訊息"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/outfit/summary": {
      "get": {
        "tags": [
          "Outfit"
        ],
        "summary": "取得場合分類統計",
        "description": "取得各個場合的穿搭總數及最近的更新日期。",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "取得場合列表成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 200
                    },
                    "status": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "取得場合列表成功"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "summaryList": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "occasionId": {
                                "type": "string",
                                "example": "campusCasual"
                              },
                              "count": {
                                "type": "integer",
                                "example": 5
                              },
                              "recentDates": {
                                "type": "array",
                                "items": {
                                  "type": "string",
                                  "example": "2024/06/01"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "身分驗證失敗 (可能原因：未提供 Token、Token 格式錯誤、Token 已過期)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 401
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "未提供 Token 或格式錯誤 / 無效的 Token 格式 / 無效的憑證或憑證已過期，請重新登入"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "伺服器錯誤",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 500
                    },
                    "status": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "內部伺服器錯誤或其他錯誤訊息"
                    },
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "example": null
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}